Reduce spinning in LockUser
LockUser today spins forever invoking the scheduler as long as
the lock is held by the bottom half(aka processor). This change
reduces the amount of spinning by strictly bounding the number
of iterations we try to acquire lock. This reduces the CPU time
in an iperf test further by a few %.

PiperOrigin-RevId: 473887032